Move more code out of legacy utils package
