+++ packages/utils/src/index.ts
@@ -1,27 +1,5 @@
 import { AssetMode, ImageExtension } from '@keystone-next/types';
 
-/**
- * Converts the first character of a string to uppercase.
- * @param {String} str The string to convert.
- * @returns The new string
- */
-export const upcase = (str: string) => str.substr(0, 1).toUpperCase() + str.substr(1);
-
-/**
- * Turns a passed in string into
- * a human readable label
- * @param {String} str The string to convert.
- * @returns The new string
- */
-export const humanize = (str: string) => {
-  return str
-    .replace(/([a-z])([A-Z]+)/g, '$1 $2')
-    .split(/\s|_|\-/)
-    .filter(i => i)
-    .map(upcase)
-    .join(' ');
-};
-
 const IMAGEREGEX = /^(local|keystone-cloud):image:([^\\\/:\n]+)\.(gif|jpg|png|webp)$/;
 const FILEREGEX = /^(local):file:([^\\\/:\n]+)/;
